requirements.txt
langchain
faiss-cpu
sentence-transformers
openai
pypdf
streamlit
